+++ R/ArchiveAsyncFrozen.R
@@ -1,7 +1,11 @@
-#' @title Rush Data Storage
+
+#' @title Frozen Rush Data Storage
 #'
 #' @description
-#' The `ArchiveAsync` stores all evaluated points and performance scores in a [rush::Rush] data base.
+#' Freezes the Redis data base of an [ArchiveAsync] to a  `data.table::data.table()`.
+#' No further points can be added to the archive but the data can be accessed and analyzed.
+#' Useful when the Redis data base is not permanently available.
+#' Use the callback [bbotk.async_freeze_archive] to freeze the archive after the optimization has finished.
 #'
 #' @section S3 Methods:
 #' * `as.data.table(archive)`\cr
@@ -9,33 +13,23 @@
 #'   Returns a tabular view of all performed function calls of the Objective.
 #'   The `x_domain` column is unnested to separate columns.
 #'
-#' @template param_search_space
-#' @template param_codomain
-#' @template param_check_values
-#' @template param_rush
-#'
-#' @template field_rush
-#'
 #' @export
-ArchiveAsync = R6Class("ArchiveAsync",
-  inherit = Archive,
+ArchiveAsyncFrozen = R6Class("ArchiveAsyncFrozen",
+  inherit = ArchiveAsync,
   cloneable = FALSE,
   public = list(
 
-    rush = NULL,
-
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
-    initialize = function(search_space, codomain, check_values = FALSE, rush) {
-      require_namespaces("rush")
-      self$rush = rush::assert_rush(rush)
-
-      super$initialize(
-        search_space = search_space,
-        codomain = codomain,
-        check_values = check_values,
-        label = "Rush Data Storage",
-        man = "bbotk::ArchiveAsync")
+    #'
+    #' @param archive ([ArchiveAsync])\cr
+    #' The archive to freeze.
+    initialize = function(archive) {
+      private$.frozen_data = copy(archive$data)
+      self$search_space = archive$search_space
+      self$codomain = archive$codomain
+      private$.label = "Frozen Data Storage"
+      private$.man = "bbotk::ArchiveAsyncFrozen"
     },
 
     #' @description
@@ -44,14 +38,13 @@ ArchiveAsync = R6Class("ArchiveAsync",
     #' @param xss (list of named `list()`)\cr
     #' List of named lists of point values.
     push_points = function(xss) {
-      if (self$check_values) map(xss, self$search_space$assert)
-      self$rush$push_tasks(xss, extra = list(list(timestamp_xs = Sys.time())))
+      stop("Archive is frozen")
     },
 
     #' @description
     #' Pop a point from the queue.
     pop_point = function() {
-      self$rush$pop_task(fields = "xs")
+      stop("Archive is frozen")
     },
 
     #' @description
@@ -62,9 +55,7 @@ ArchiveAsync = R6Class("ArchiveAsync",
     #' @param extra (`list()`)\cr
     #' Named list of additional information.
     push_running_point = function(xs, extra = NULL) {
-      if (self$check_values) self$search_space$assert(xs)
-      extra = c(list(timestamp_xs = Sys.time()), extra)
-      self$rush$push_running_tasks(list(xs), extra = list(extra))
+      stop("Archive is frozen")
     },
 
     #' @description
@@ -79,8 +70,7 @@ ArchiveAsync = R6Class("ArchiveAsync",
     #' @param extra (`list()`)\cr
     #' Named list of additional information.
     push_result = function(key, ys, x_domain, extra = NULL) {
-      extra = c(list(x_domain = list(x_domain), timestamp_ys = Sys.time()), extra)
-      self$rush$push_results(key, list(ys), extra = list(extra))
+      stop("Archive is frozen")
     },
 
     #' @description
@@ -91,7 +81,7 @@ ArchiveAsync = R6Class("ArchiveAsync",
     #' @param message (`character()`)\cr
     #' Error message.
     push_failed_point = function(key, message) {
-      self$rush$push_failed(key, list(list(message = message)))
+      stop("Archive is frozen")
     },
 
     #' @description
@@ -110,70 +100,18 @@ ArchiveAsync = R6Class("ArchiveAsync",
       states = c("queued", "running", "finished", "failed"),
       reset_cache = FALSE
       ) {
-      self$rush$fetch_tasks_with_state(fields, states, reset_cache)
-    },
-
-    #' @description
-    #' Returns the best scoring evaluation(s).
-    #' For single-crit optimization, the solution that minimizes / maximizes the objective function.
-    #' For multi-crit optimization, the Pareto set / front.
-    #'
-    #' @param n_select (`integer(1L)`)\cr
-    #' Amount of points to select.
-    #' Ignored for multi-crit optimization.
-    #' @param ties_method (`character(1L)`)\cr
-    #' Method to break ties when multiple points have the same score.
-    #' Either `"first"` (default) or `"random"`.
-    #' Ignored for multi-crit optimization.
-    #' If `n_select > 1L`, the tie method is ignored and the first point is returned.
-    #'
-    #' @return [data.table::data.table()]
-    best = function(n_select = 1, ties_method = "first") {
-      assert_count(n_select)
-      tab = self$finished_data
-
-      if (self$codomain$target_length == 1L) {
-        if (n_select == 1L) {
-          # use which_max to find the best point
-          y = tab[[self$cols_y]] * -self$codomain$direction
-          ii = which_max(y, ties_method = ties_method)
-          tab[ii]
-        } else {
-          # use data.table fast sort to find the best points
-          setorderv(tab, cols = self$cols_y, order = self$codomain$direction)
-          head(tab, n_select)
-        }
-      } else {
-        # use non-dominated sorting to find the best points
-        ymat = t(as.matrix(tab[, self$cols_y, with = FALSE]))
-        ymat = self$codomain$direction * ymat
-        tab[!is_dominated(ymat)]
-      }
-    },
-
-    #' @description
-    #' Calculate best points w.r.t. non dominated sorting with hypervolume contribution.
-    #'
-    #' @template param_n_select
-    #' @template param_ref_point
-    #'
-    #' @return [data.table::data.table()]
-    nds_selection = function(n_select = 1, ref_point = NULL) {
-      tab = self$finished_data
-      assert_int(n_select, lower = 1L, upper = nrow(tab))
-
-      points = t(as.matrix(tab[, self$cols_y, with = FALSE]))
-      minimize = map_lgl(self$codomain$target_tags, has_element, "minimize")
-      inds = nds_selection(points, n_select, ref_point, minimize)
-      tab[inds, ]
+      stop("Archive is frozen")
     },
 
     #' @description
     #' Clear all evaluation results from archive.
     clear = function() {
-      self$rush$reset()
-      super$clear()
+      stop("Archive is frozen")
     }
+  ),
+
+  private = list(
+    .frozen_data = NULL
   ),
 
   active = list(
@@ -182,68 +120,68 @@ ArchiveAsync = R6Class("ArchiveAsync",
     #' Data table with all finished points.
     data = function(rhs) {
       assert_ro_binding(rhs)
-      self$data_with_state()
+      private$.frozen_data
     },
 
     #' @field queued_data ([data.table::data.table])\cr
     #' Data table with all queued points.
     queued_data = function() {
-      self$rush$fetch_queued_tasks()
+      self$data["queued", , on = "state"]
     },
 
     #' @field running_data ([data.table::data.table])\cr
     #' Data table with all running points.
     running_data = function() {
-      self$rush$fetch_running_tasks()
+      self$data["running", , on = "state"]
     },
 
     #' @field finished_data ([data.table::data.table])\cr
     #' Data table with all finished points.
     finished_data = function() {
-      self$rush$fetch_finished_tasks()
+      self$data["finished", , on = "state"]
     },
 
     #' @field failed_data ([data.table::data.table])\cr
     #' Data table with all failed points.
     failed_data = function() {
-      self$rush$fetch_failed_tasks()
+      self$data["failed", , on = "state"]
     },
 
     #' @field n_queued (`integer(1)`)\cr
     #' Number of queued points.
     n_queued = function() {
-      self$rush$n_queued_tasks
+      nrow(self$queued_data)
     },
 
     #' @field n_running (`integer(1)`)\cr
     #' Number of running points.
     n_running = function() {
-      self$rush$n_running_tasks
+      nrow(self$running_data)
     },
 
     #' @field n_finished (`integer(1)`)\cr
     #' Number of finished points.
     n_finished = function() {
-      self$rush$n_finished_tasks
+      nrow(self$finished_data)
     },
 
     #' @field n_failed (`integer(1)`)\cr
     #' Number of failed points.
     n_failed = function() {
-      self$rush$n_failed_tasks
+      nrow(self$failed_data)
     },
 
     #' @field n_evals (`integer(1)`)\cr
     #' Number of evaluations stored in the archive.
     n_evals = function() {
-      self$rush$n_finished_tasks + self$rush$n_failed_tasks
+      nrow(self$finished_data) + nrow(self$failed_data)
     }
   )
 )
 
 #' @export
 as.data.table.ArchiveAsync = function(x, keep.rownames = FALSE, unnest = "x_domain", ...) { # nolint
-  data = x$data_with_state()
+  data = x$data
   cols = intersect(unnest, names(data))
   unnest(data, cols, prefix = "{col}_")
 }