+++ R/OptimInstanceAsyncSingleCrit.R
@@ -52,18 +52,31 @@ OptimInstanceAsyncSingleCrit = R6Class("OptimInstanceAsyncSingleCrit",
     #'
     #' @param y (`numeric(1)`)\cr
     #' Optimal outcome.
+    #' @param extra (`data.table::data.table()`)\cr
+    #' Additional information.
     #' @param ... (`any`)\cr
     #' ignored.
-    assign_result = function(xdt, y, ...) {
-      # FIXME: We could have one way that just lets us put a 1xn DT as result directly.
-      assert_data_table(xdt)
-      assert_names(names(xdt), must.include = self$search_space$ids())
-      assert_number(y)
-      assert_names(names(y), permutation.of = self$objective$codomain$target_ids)
-      x_domain = unlist(transform_xdt_to_xss(xdt, self$search_space), recursive = FALSE)
+    assign_result = function(xdt, y, extra = NULL, ...) {
+      # assign for callbacks
+      private$.result_xdt = xdt
+      private$.result_y = y
+      private$.result_extra = extra
+
+      call_back("on_result_begin", self$objective$callbacks, self$objective$context)
+
+      # assert inputs
+      assert_names(names(private$.result_xdt), must.include = self$search_space$ids())
+      assert_data_table(private$.result_xdt)
+      assert_number(private$.result_y)
+      assert_names(names(private$.result_y), permutation.of = self$objective$codomain$target_ids)
+      assert_data_table(private$.result_extra, null.ok = TRUE)
+
+      # add x_domain to result
+      x_domain = unlist(transform_xdt_to_xss(private$.result_xdt, self$search_space), recursive = FALSE)
       if (is.null(x_domain)) x_domain = list()
-      private$.result = cbind(xdt, x_domain = list(x_domain), t(y)) # t(y) so the name of y stays
-      call_back("on_result", self$objective$callbacks, self$objective$context)
+
+      private$.result = cbind(private$.result_xdt, x_domain = list(x_domain), t(private$.result_y)) # t(y) so the name of y stays
+      call_back("on_result_end", self$objective$callbacks, self$objective$context)
     }
   ),
 
@@ -80,5 +93,10 @@ OptimInstanceAsyncSingleCrit = R6Class("OptimInstanceAsyncSingleCrit",
     result_y = function() {
       unlist(private$.result[, self$objective$codomain$ids(), with = FALSE])
     }
+  ),
+
+  private = list(
+    # intermediate objects
+    .result_y = NULL
   )
 )