+++ R/OptimInstanceBatchMultiCrit.R
@@ -44,18 +44,32 @@ OptimInstanceBatchMultiCrit = R6Class("OptimInstanceBatchMultiCrit",
     #' The [Optimizer] object writes the best found points
     #' and estimated performance values here (probably the Pareto set / front).
     #' For internal use.
+    #'
+    #' @param extra (`data.table::data.table()`)\cr
+    #' Additional information.
     #' @param ... (`any`)\cr
     #' ignored.
-    assign_result = function(xdt, ydt, ...) {
-      # FIXME: We could have one way that just lets us put a 1xn DT as result directly.
-      assert_data_table(xdt)
-      assert_names(names(xdt), must.include = self$search_space$ids())
-      assert_data_table(ydt)
-      assert_names(names(ydt), permutation.of = self$objective$codomain$ids())
-      x_domain = transform_xdt_to_xss(xdt, self$search_space)
+    assign_result = function(xdt, ydt, extra = NULL, ...) {
+      # assign for callbacks
+      private$.result_xdt = xdt
+      private$.result_ydt = ydt
+      private$.result_extra = extra
+
+      call_back("on_result_begin", self$objective$callbacks, self$objective$context)
+
+      # assert inputs
+      assert_data_table(private$.result_xdt)
+      assert_names(names(private$.result_xdt), must.include = self$search_space$ids())
+      assert_data_table(private$.result_ydt)
+      assert_names(names(private$.result_ydt), permutation.of = self$objective$codomain$ids())
+      assert_data_table(private$.result_extra, null.ok = TRUE)
+
+      # add x_domain to result
+      x_domain = transform_xdt_to_xss(private$.result_xdt, self$search_space)
       if (length(x_domain) == 0) x_domain = list(list())
-      private$.result = cbind(xdt, x_domain = x_domain, ydt)
-      call_back("on_result", self$objective$callbacks, self$objective$context)
+
+      private$.result = cbind(private$.result_xdt, x_domain = x_domain, private$.result_ydt)
+      call_back("on_result_end", self$objective$callbacks, self$objective$context)
     }
   ),
 
@@ -71,5 +85,10 @@ OptimInstanceBatchMultiCrit = R6Class("OptimInstanceBatchMultiCrit",
     result_y = function() {
       private$.result[, self$objective$codomain$ids(), with = FALSE]
     }
+  ),
+
+  private = list(
+    # intermediate objects
+    .result_ydt = NULL
   )
 )