+++ R/OptimizerAsync.R
@@ -109,8 +109,7 @@ optimize_async_default = function(instance, optimizer, design = NULL, n_workers 
         worker_loop = bbotk_worker_loop,
         packages = c(optimizer$packages, "bbotk"), # add packages from objective
         optimizer = optimizer,
-        instance = instance,
-        wait_for_workers = TRUE)
+        instance = instance)
     } else {
        stop("No rush plan available to start local workers and no pre-started remote workers found. See `?rush::rush_plan()`.")
     }
@@ -141,7 +140,7 @@ optimize_async_default = function(instance, optimizer, design = NULL, n_workers 
   get_private(optimizer)$.assign_result(instance)
   lg$info("Finished optimizing after %i evaluation(s)", instance$archive$n_evals)
   lg$info("Result:")
-  lg$info(capture.output(print(instance$result, lass = FALSE, row.names = FALSE, print.keys = FALSE)))
+  lg$info(capture.output(print(instance$result, class = FALSE, row.names = FALSE, print.keys = FALSE)))
 
   call_back("on_optimization_end", instance$objective$callbacks, instance$objective$context)
   return(instance$result)